+++ Package.swift
@@ -12,8 +12,8 @@ let package = Package(
     targets: [
         .binaryTarget(
           name: "FuturaeKit",
-          url: "https://artifactory.futurae.com/artifactory/futurae-ios/FuturaeKit-v1.2.0.xcframework.zip",
-          checksum: "69c6e9b177bf2f1f9ced27695cc69009ecc60dffbb82022ab528df023dcb73b3"
+          url: "https://artifactory.futurae.com/artifactory/futurae-ios/FuturaeKit-v1.3.0.xcframework.zip",
+          checksum: "d577712d860bf7d79c00240b006d5106a01749a9ddf82b43330fc298118b87ce"
         )
     ]
 )